@@ -4,7 +4,7 @@ version: 1.2.0
 author: Harpreet Seera <harpreetseera8525@gmail.com>
 homepage: https://github.com/hd-motion/flutter_imageview_360
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
