@@ -30,11 +30,11 @@ class ImageView360 extends StatefulWidget {
   final RotationDirection rotationDirection;
 
   // Callback function to provide you the index of current image when image frame is changed.
-  final Function(int currentImageIndex) onImageIndexChanged;
+  final Function(int currentImageIndex)? onImageIndexChanged;
 
   ImageView360({
-    @required Key key,
-    @required this.imageList,
+    required Key key,
+    required this.imageList,
     this.autoRotate = false,
     this.allowSwipeToRotate = true,
     this.rotationCount = 1,
@@ -49,10 +49,10 @@ class ImageView360 extends StatefulWidget {
 }
 
 class _ImageView360State extends State<ImageView360> {
-  int rotationIndex, senstivity;
+  late int rotationIndex, senstivity;
   int rotationCompleted = 0;
   double localPosition = 0.0;
-  Function(int currentImageIndex) onImageIndexChanged;
+  Function(int currentImageIndex)? onImageIndexChanged;
 
   @override
   void initState() {
@@ -128,7 +128,9 @@ class _ImageView360State extends State<ImageView360> {
           }
         });
 
-        onImageIndexChanged(rotationIndex);
+        if (onImageIndexChanged != null) {
+          onImageIndexChanged!(rotationIndex);
+        }
       }
       //Recursive call
       rotateImage();
@@ -152,7 +154,9 @@ class _ImageView360State extends State<ImageView360> {
           rotationIndex = 0;
         }
       });
-      onImageIndexChanged(rotationIndex);
+      if (onImageIndexChanged != null) {
+        onImageIndexChanged!(rotationIndex);
+      }
     }
   }
 
@@ -176,7 +180,9 @@ class _ImageView360State extends State<ImageView360> {
           rotationIndex = widget.imageList.length - 1;
         }
       });
-      onImageIndexChanged(rotationIndex);
+      if (onImageIndexChanged != null) {
+        onImageIndexChanged!(rotationIndex);
+      }
     }
   }
 }
